+++ Linux/src/OSUtils.cpp
@@ -32,3 +32,44 @@ bool OSUtils::deleteFileIfExists(const char* filename)
     // Return true if the file was deleted or if it did not exist
     return true;
 }
+
+
+bool OSUtils::appendToFile(std::string fileName, std::string content)
+{
+    std::ofstream outFile(fileName, std::ios_base::app);
+    if (!outFile) 
+    {
+        // If the file doesn't exist, create a new one
+        outFile.open(fileName);
+        if (!outFile) 
+        {
+            std::cerr << "Error creating file: " << fileName << std::endl;
+            return false;
+        }
+    }
+
+    outFile << content;
+    outFile.close();
+    return true;
+}
+
+std::string OSUtils::getFileName(const char *fullPath, bool removeExtension)
+{
+    std::string fileName(fullPath);
+    size_t lastSlash = fileName.find_last_of("/\\");
+    if (lastSlash != std::string::npos) {
+        fileName = fileName.substr(lastSlash + 1);
+    }
+
+    if (removeExtension) 
+    {
+        size_t lastDot = fileName.find_last_of(".");
+        if (lastDot != std::string::npos) {
+            fileName = fileName.substr(0, lastDot);
+        }
+    }
+
+    return fileName;
+}
+
+